fix unittest script

+++ unittest.sh
@@ -1,7 +1,19 @@
 #!/bin/bash
 
+pkg=$1
+if [ -z "$pkg" ]; then
+  echo "Usage $0 <frontend|backend>"
+  exit 1
+fi
+
 # Cleanup
 rm -rf .bundle vendor *.deb
+
+if [ "$pkg"="frontend" ]; then
+  exit 0
+elif [ "$pkg"="backend" ]; then
+  SPECS="spec/app/models/ spec/framework/persistence/memory_spec.rb"
+fi
 
 # Install all dependencies locally
 bundle install --deployment
@@ -10,5 +22,4 @@ bundle install --deployment
 bundle exec rspec \
 	--format RspecJunitFormatter \
 	--out rspec.xml \
-	spec/app/models/ \
-	spec/framework/persistence/memory_spec.rb
+	$SPECS
